+++ CHANGELOG.md
@@ -9,6 +9,8 @@ is based on https://keepachangelog.com/en/1.1.0/.
 
 - `alloc` feature for enabling types, impls and fns that require the `alloc`
   crate.
+- Basic support for 48bit and 64bit long instructions.
+- The `instruction::base::Set` associated fns `decode_48` and `decode_64`.
 - A trait `decoder::unit::DebugIOptions` combining `decoder::unit::IOptions` and
   `Debug`.
 - A `decoder::unit::Unit` type `decoder::unit::PULP` for PULP's `rv_tracer`.
@@ -55,6 +57,8 @@ is based on https://keepachangelog.com/en/1.1.0/.
 
 ### Changed
 
+- The `instruction::Bits` has two more variants `Bit48` and `Bit64`.
+- The `instruction::Size` has two more variants `Wide` and `ExtraWide`.
 - `(A, B)` now implements `binary::Binary` only if both binaries agree on their
   `binary::Binary::Error` type, and selects `B` only if `A` yield a "miss".
 - The `tracer::Tracer` fns `process_te_inst`, `process_sync` and